land-encroachment-project/backend/requirements.txt
Flask
Flask-Cors
Flask-SocketIO
eventlet
gunicorn
python-dotenv
opencv-python
numpy
Werkzeug
PyJWT
shapely
geojson
reportlab
APScheduler
pytesseract
pypdf
